refactor: Merge upstream changes and fix test parameter calls
