+++ zk/conn_test.go
@@ -1,28 +1,16 @@
 package zk
 
 import (
+	"context"
+	"fmt"
 	"io/ioutil"
+	"sync"
 	"testing"
 	"time"
 )
 
-func TestRecurringReAuthHang(t *testing.T) {
-	t.Skip("Race condition in test")
-
-	sessionTimeout := 2 * time.Second
-
-	finish := make(chan struct{})
-	defer close(finish)
-	go func() {
-		select {
-		case <-finish:
-			return
-		case <-time.After(5 * sessionTimeout):
-			panic("expected not hang")
-		}
-	}()
-
-	zkC, err := StartTestCluster(2, ioutil.Discard, ioutil.Discard)
+func TestIntegration_RecurringReAuthHang(t *testing.T) {
+	zkC, err := StartTestCluster(t, 3, ioutil.Discard, ioutil.Discard)
 	if err != nil {
 		panic(err)
 	}
@@ -32,28 +20,71 @@ func TestRecurringReAuthHang(t *testing.T) {
 	if err != nil {
 		panic(err)
 	}
-	for conn.State() != StateHasSession {
-		time.Sleep(50 * time.Millisecond)
-	}
+	defer conn.Close()
 
-	go func() {
-		for range evtC {
-		}
-	}()
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
+	defer cancel()
 
+	waitForSession(ctx, evtC)
 	// Add auth.
 	conn.AddAuth("digest", []byte("test:test"))
 
-	currentServer := conn.Server()
-	conn.debugCloseRecvLoop = true
-	conn.debugReauthDone = make(chan struct{})
-	zkC.StopServer(currentServer)
-	// wait connect to new zookeeper.
-	for conn.Server() == currentServer && conn.State() != StateHasSession {
-		time.Sleep(100 * time.Millisecond)
+	var reauthCloseOnce sync.Once
+	reauthSig := make(chan struct{}, 1)
+	conn.resendZkAuthFn = func(ctx context.Context, c *Conn) error {
+		// in current implimentation the reauth might be called more than once based on various conditions
+		reauthCloseOnce.Do(func() { close(reauthSig) })
+		return resendZkAuth(ctx, c)
 	}
 
-	<-conn.debugReauthDone
+	conn.debugCloseRecvLoop = true
+	currentServer := conn.Server()
+	zkC.StopServer(currentServer)
+	// wait connect to new zookeeper.
+	ctx, cancel = context.WithTimeout(context.Background(), time.Second*5)
+	defer cancel()
+
+	waitForSession(ctx, evtC)
+
+	select {
+	case _, ok := <-reauthSig:
+		if !ok {
+			return // we closed the channel as expected
+		}
+		t.Fatal("reauth testing channel should have been closed")
+	case <-ctx.Done():
+		t.Fatal(ctx.Err())
+	}
+}
+
+func TestConcurrentReadAndClose(t *testing.T) {
+	WithListenServer(t, func(server string) {
+		conn, _, err := Connect([]string{server}, 15*time.Second)
+		if err != nil {
+			t.Fatalf("Failed to create Connection %s", err)
+		}
+
+		okChan := make(chan struct{})
+		var setErr error
+		go func() {
+			_, setErr = conn.Create("/test-path", []byte("test data"), 0, WorldACL(PermAll))
+			close(okChan)
+		}()
+
+		go func() {
+			time.Sleep(1 * time.Second)
+			conn.Close()
+		}()
+
+		select {
+		case <-okChan:
+			if setErr != ErrConnectionClosed {
+				t.Fatalf("unexpected error returned from Set %v", setErr)
+			}
+		case <-time.After(3 * time.Second):
+			t.Fatal("apparent deadlock!")
+		}
+	})
 }
 
 func TestDeadlockInClose(t *testing.T) {
@@ -78,5 +109,90 @@ func TestDeadlockInClose(t *testing.T) {
 	case <-okChan:
 	case <-time.After(3 * time.Second):
 		t.Fatal("apparent deadlock!")
+	}
+}
+
+func TestNotifyWatches(t *testing.T) {
+	cases := []struct {
+		eType   EventType
+		path    string
+		watches map[WatchPathType]bool
+	}{
+		{
+			EventNodeCreated, "/",
+			map[WatchPathType]bool{
+				{"/", WatchTypeExist}: true,
+				{"/", WatchTypeChild}: false,
+				{"/", WatchTypeData}:  false,
+			},
+		},
+		{
+			EventNodeCreated, "/a",
+			map[WatchPathType]bool{
+				{"/b", WatchTypeExist}: false,
+			},
+		},
+		{
+			EventNodeDataChanged, "/",
+			map[WatchPathType]bool{
+				{"/", WatchTypeExist}: true,
+				{"/", WatchTypeData}:  true,
+				{"/", WatchTypeChild}: false,
+			},
+		},
+		{
+			EventNodeChildrenChanged, "/",
+			map[WatchPathType]bool{
+				{"/", WatchTypeExist}: false,
+				{"/", WatchTypeData}:  false,
+				{"/", WatchTypeChild}: true,
+			},
+		},
+		{
+			EventNodeDeleted, "/",
+			map[WatchPathType]bool{
+				{"/", WatchTypeExist}: true,
+				{"/", WatchTypeData}:  true,
+				{"/", WatchTypeChild}: true,
+			},
+		},
+	}
+
+	conn := &Conn{watchers: make(map[WatchPathType]map[*Watcher]bool)}
+
+	for idx, c := range cases {
+		t.Run(fmt.Sprintf("#%d %s", idx, c.eType), func(t *testing.T) {
+			c := c
+
+			notifications := make([]struct {
+				path   string
+				notify bool
+				ch     <-chan Event
+			}, len(c.watches))
+
+			var idx int
+			for wpt, expectEvent := range c.watches {
+				ch := conn.addWatcher(wpt.Path, wpt.WType).EvtCh
+				notifications[idx].path = wpt.Path
+				notifications[idx].notify = expectEvent
+				notifications[idx].ch = ch
+				idx++
+			}
+			ev := Event{Type: c.eType, Path: c.path}
+			conn.notifyWatches(ev)
+
+			for _, res := range notifications {
+				select {
+				case e := <-res.ch:
+					if !res.notify || e.Path != res.path {
+						t.Fatal("unexpeted notification received")
+					}
+				default:
+					if res.notify {
+						t.Fatal("expected notification not received")
+					}
+				}
+			}
+		})
 	}
 }
