+++ zk/lock_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestIntegration_Lock(t *testing.T) {
-	ts, err := StartTestCluster(t, 1, nil, logWriter{t: t, p: "[ZKERR] "})
+	ts, err := StartTestCluster(t, 1, logWriter{t: t, p: "[ZKERR] "}, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -64,7 +64,7 @@ func TestIntegration_Lock(t *testing.T) {
 // This tests creating a lock with a path that's more than 1 node deep (e.g. "/test-multi-level/lock"),
 // when a part of that path already exists (i.e. "/test-multi-level" node already exists).
 func TestIntegration_MultiLevelLock(t *testing.T) {
-	ts, err := StartTestCluster(t, 1, nil, logWriter{t: t, p: "[ZKERR] "})
+	ts, err := StartTestCluster(t, 1, logWriter{t: t, p: "[ZKERR] "}, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
