+++ zk/server_java.go
@@ -1,11 +1,13 @@
 package zk
 
 import (
+	"context"
 	"fmt"
 	"io"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"testing"
 )
 
 type ErrMissingServerConfigField string
@@ -15,13 +17,65 @@ func (e ErrMissingServerConfigField) Error() string {
 }
 
 const (
-	DefaultServerTickTime                 = 2000
+	DefaultServerTickTime                 = 500
 	DefaultServerInitLimit                = 10
 	DefaultServerSyncLimit                = 5
 	DefaultServerAutoPurgeSnapRetainCount = 3
 	DefaultPeerPort                       = 2888
 	DefaultLeaderElectionPort             = 3888
 )
+
+type server struct {
+	stdout, stderr io.Writer
+	cmdString      string
+	cmdArgs        []string
+	cmdEnv         []string
+	cmd            *exec.Cmd
+	// this cancel will kill the command being run in this case the server itself.
+	cancelFunc context.CancelFunc
+}
+
+func NewIntegrationTestServer(t *testing.T, configPath string, stdout, stderr io.Writer) (*server, error) {
+	// allow external systems to configure this zk server bin path.
+	zkPath := os.Getenv("ZOOKEEPER_BIN_PATH")
+	if zkPath == "" {
+		// default to a static reletive path that can be setup with a build system
+		zkPath = "zookeeper/bin"
+	}
+	if _, err := os.Stat(zkPath); err != nil {
+		if os.IsNotExist(err) {
+			return nil, fmt.Errorf("zk: could not find testing zookeeper bin path at %q: %v ", zkPath, err)
+		}
+	}
+	// password is 'test'
+	superString := `SERVER_JVMFLAGS=-Dzookeeper.DigestAuthenticationProvider.superDigest=super:D/InIHSb7yEEbrWz8b9l71RjZJU=`
+	// enable TTL
+	superString += ` -Dzookeeper.extendedTypesEnabled=true -Dzookeeper.emulate353TTLNodes=true`
+
+	return &server{
+		cmdString: filepath.Join(zkPath, "zkServer.sh"),
+		cmdArgs:   []string{"start-foreground", configPath},
+		cmdEnv:    []string{superString},
+		stdout:    stdout, stderr: stderr,
+	}, nil
+}
+
+func (srv *server) Start() error {
+	ctx, cancel := context.WithCancel(context.Background())
+	srv.cancelFunc = cancel
+
+	srv.cmd = exec.CommandContext(ctx, srv.cmdString, srv.cmdArgs...)
+	srv.cmd.Stdout = srv.stdout
+	srv.cmd.Stderr = srv.stderr
+	srv.cmd.Env = append(os.Environ(), srv.cmdEnv...)
+
+	return srv.cmd.Start()
+}
+
+func (srv *server) Stop() error {
+	srv.cancelFunc()
+	return srv.cmd.Wait()
+}
 
 type ServerConfigServer struct {
 	ID                 int
@@ -42,6 +96,9 @@ type ServerConfig struct {
 }
 
 func (sc ServerConfig) Marshall(w io.Writer) error {
+	// the admin server is not wanted in test cases as it slows the startup process and is
+	// of little unit test value.
+	fmt.Fprintln(w, "admin.enableServer=false")
 	if sc.DataDir == "" {
 		return ErrMissingServerConfigField("dataDir")
 	}
@@ -69,68 +126,46 @@ func (sc ServerConfig) Marshall(w io.Writer) error {
 		fmt.Fprintf(w, "autopurge.snapRetainCount=%d\n", sc.AutoPurgeSnapRetainCount)
 		fmt.Fprintf(w, "autopurge.purgeInterval=%d\n", sc.AutoPurgePurgeInterval)
 	}
-	if len(sc.Servers) > 0 {
-		for _, srv := range sc.Servers {
-			if srv.PeerPort <= 0 {
-				srv.PeerPort = DefaultPeerPort
-			}
-			if srv.LeaderElectionPort <= 0 {
-				srv.LeaderElectionPort = DefaultLeaderElectionPort
-			}
-			fmt.Fprintf(w, "server.%d=%s:%d:%d\n", srv.ID, srv.Host, srv.PeerPort, srv.LeaderElectionPort)
+	// enable reconfig.
+	// TODO: allow setting this
+	fmt.Fprintln(w, "reconfigEnabled=true")
+	fmt.Fprintln(w, "4lw.commands.whitelist=*")
+
+	if len(sc.Servers) < 2 {
+		// if we dont have more than 2 servers we just dont specify server list to start in standalone mode
+		// see https://zookeeper.apache.org/doc/current/zookeeperStarted.html#sc_InstallingSingleMode for more details.
+		return nil
+	}
+	// if we then have more than one server force it to be distributed
+	fmt.Fprintln(w, "standaloneEnabled=false")
+
+	for _, srv := range sc.Servers {
+		if srv.PeerPort <= 0 {
+			srv.PeerPort = DefaultPeerPort
 		}
+		if srv.LeaderElectionPort <= 0 {
+			srv.LeaderElectionPort = DefaultLeaderElectionPort
+		}
+		fmt.Fprintf(w, "server.%d=%s:%d:%d\n", srv.ID, srv.Host, srv.PeerPort, srv.LeaderElectionPort)
 	}
 	return nil
 }
 
-var jarSearchPaths = []string{
-	"zookeeper-*/contrib/fatjar/zookeeper-*-fatjar.jar",
-	"../zookeeper-*/contrib/fatjar/zookeeper-*-fatjar.jar",
-	"/usr/share/java/zookeeper-*.jar",
-	"/usr/local/zookeeper-*/contrib/fatjar/zookeeper-*-fatjar.jar",
-	"/usr/local/Cellar/zookeeper/*/libexec/contrib/fatjar/zookeeper-*-fatjar.jar",
-}
-
-func findZookeeperFatJar() string {
-	var paths []string
-	zkPath := os.Getenv("ZOOKEEPER_PATH")
-	if zkPath == "" {
-		paths = jarSearchPaths
-	} else {
-		paths = []string{filepath.Join(zkPath, "contrib/fatjar/zookeeper-*-fatjar.jar")}
-	}
-	for _, path := range paths {
-		matches, _ := filepath.Glob(path)
-		// TODO: could sort by version and pick latest
-		if len(matches) > 0 {
-			return matches[0]
+// this is a helper to wait for the zk connection to at least get to the HasSession state
+func waitForSession(ctx context.Context, eventChan <-chan Event) error {
+	select {
+	case event, ok := <-eventChan:
+		// The eventChan is used solely to determine when the ZK conn has
+		// stopped.
+		if !ok {
+			return fmt.Errorf("connection closed before state reached")
 		}
-	}
-	return ""
-}
-
-type Server struct {
-	JarPath        string
-	ConfigPath     string
-	Stdout, Stderr io.Writer
-
-	cmd *exec.Cmd
-}
-
-func (srv *Server) Start() error {
-	if srv.JarPath == "" {
-		srv.JarPath = findZookeeperFatJar()
-		if srv.JarPath == "" {
-			return fmt.Errorf("zk: unable to find server jar")
+		if event.State == StateHasSession {
+			return nil
 		}
+	case <-ctx.Done():
+		return ctx.Err()
 	}
-	srv.cmd = exec.Command("java", "-jar", srv.JarPath, "server", srv.ConfigPath)
-	srv.cmd.Stdout = srv.Stdout
-	srv.cmd.Stderr = srv.Stderr
-	return srv.cmd.Start()
-}
 
-func (srv *Server) Stop() error {
-	srv.cmd.Process.Signal(os.Kill)
-	return srv.cmd.Wait()
+	return nil
 }
